+++ iceberg/iceberg-rest-server/src/main/java/org/apache/gravitino/listener/api/event/IcebergDropTableFailureEvent.java
@@ -24,7 +24,7 @@ import org.apache.gravitino.annotation.DeveloperApi;
 
 /** Represent a failure event when dropping Iceberg table failed. */
 @DeveloperApi
-public class IcebergDropTableFailureEvent extends IcebergRESTFailureEvent {
+public class IcebergDropTableFailureEvent extends IcebergTableFailureEvent {
   private boolean purgeRequested;
 
   public IcebergDropTableFailureEvent(
