+++ iceberg/iceberg-rest-server/src/main/java/org/apache/gravitino/listener/api/event/IcebergTableExistsFailureEvent.java
@@ -24,7 +24,7 @@ import org.apache.gravitino.annotation.DeveloperApi;
 
 /** Represent a failure event when check Iceberg table exist failed. */
 @DeveloperApi
-public class IcebergTableExistsFailureEvent extends IcebergRESTFailureEvent {
+public class IcebergTableExistsFailureEvent extends IcebergTableFailureEvent {
   public IcebergTableExistsFailureEvent(String user, NameIdentifier nameIdentifier, Exception e) {
     super(user, nameIdentifier, e);
   }
